split logs to separate file
